+++ src/infra/database/prisma/mappers/prismaNotificationMapper.ts
@@ -1,4 +1,6 @@
+import { Notification as RawNotification } from '@prisma/client';
 import { Notification } from '@application/entities/notification';
+import { Content } from '@application/entities/content';
 
 export class PrismaNotificationMapper {
   static toPrisma(notification: Notification) {
@@ -6,9 +8,23 @@ export class PrismaNotificationMapper {
       id: notification.id,
       category: notification.category,
       content: notification.content.value,
-      recepientId: notification.recipientId,
+      recipientId: notification.recipientId,
       readAt: notification.readAt,
       createdAt: notification.createdAt,
     };
+  }
+
+  static toDomain(raw: RawNotification): Notification {
+    return new Notification(
+      {
+        category: raw.content,
+        content: new Content(raw.content),
+        recipientId: raw.recipientId,
+        readAt: raw.readAt,
+        canceledAt: raw.canceledAt,
+        createdAt: raw.createdAt,
+      },
+      raw.id,
+    );
   }
 }