+++ src/infra/http/http.module.ts
@@ -1,3 +1,8 @@
+import { CancelNotification } from '@application/useCases/cancelNotification';
+import { CountRecipientNotifications } from '@application/useCases/countRecipientNotificationsMyWay';
+import { GetRecipientNotifications } from '@application/useCases/getRecipientNotifications';
+import { ReadNotification } from '@application/useCases/readNotification';
+import { UnreadNotification } from '@application/useCases/unreadNotification';
 import { Module } from '@nestjs/common';
 import { SendNotification } from 'src/application/useCases/sendNotification';
 import { DatabaseModule } from '../database/prisma/database.module';
@@ -6,6 +11,13 @@ import { NotificationsController } from './controllers/notifications.controller'
 @Module({
   imports: [DatabaseModule],
   controllers: [NotificationsController],
-  providers: [SendNotification],
+  providers: [
+    CancelNotification,
+    CountRecipientNotifications,
+    GetRecipientNotifications,
+    ReadNotification,
+    UnreadNotification,
+    SendNotification,
+  ],
 })
 export class HttpModule {}